🏗 按non-web和web-mvc拆分slardar

+++ wings/slardar/src/main/java/pro/fessional/wings/slardar/httprest/OkHttpClientHelper.java
@@ -14,9 +14,6 @@ import okhttp3.ResponseBody;
 import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
-import org.springframework.http.HttpMethod;
-import org.springframework.http.client.ClientHttpRequestFactory;
-import org.springframework.http.client.OkHttp3ClientHttpRequestFactory;
 import pro.fessional.mirana.data.Null;
 import pro.fessional.mirana.io.InputStreams;
 import pro.fessional.mirana.netx.SslTrustAll;
@@ -164,32 +161,29 @@ public class OkHttpClientHelper {
     }
 
 
-    public static byte @NotNull[] download(OkHttpClient client, String url) {
-        return download(client, url, HttpMethod.GET);
+    public static byte @NotNull [] download(OkHttpClient client, String url) {
+        return download(client, url, "GET");
     }
 
-    public static byte @NotNull[] download(OkHttpClient client, String url, HttpMethod method) {
+    public static byte @NotNull [] download(OkHttpClient client, String url, String method) {
         Request.Builder builder = new Request.Builder().url(url);
-        switch (method) {
-            case GET:
-                builder.get();
-                break;
-            case POST:
-                builder.post(EMPTY);
-                break;
-            case PUT:
-                builder.put(EMPTY);
-                break;
-            case HEAD:
-                builder.head();
-                break;
-            case PATCH:
-                builder.patch(EMPTY);
-                break;
-            case DELETE:
-                builder.delete();
-                break;
-            default:
+        if ("GET".equalsIgnoreCase(method)) {
+            builder.get();
+        }
+        else if ("POST".equalsIgnoreCase(method)) {
+            builder.post(EMPTY);
+        }
+        else if ("PUT".equalsIgnoreCase(method)) {
+            builder.put(EMPTY);
+        }
+        else if ("HEAD".equalsIgnoreCase(method)) {
+            builder.head();
+        }
+        else if ("PATCH".equalsIgnoreCase(method)) {
+            builder.patch(EMPTY);
+        }
+        else if ("DELETE".equalsIgnoreCase(method)) {
+            builder.delete();
         }
 
         try (Response response = client.newCall(builder.build()).execute()) {
@@ -237,10 +231,6 @@ public class OkHttpClientHelper {
     public static void sslTrustAll(OkHttpClient.Builder builder) {
         builder.sslSocketFactory(SslTrustAll.SSL_SOCKET_FACTORY, SslTrustAll.X509_TRUST_MANAGER)
                .hostnameVerifier(SslTrustAll.HOSTNAME_VERIFIER);
-    }
-
-    public static ClientHttpRequestFactory requestFactory(OkHttpClient client) {
-        return new OkHttp3ClientHttpRequestFactory(client);
     }
 
     public static void hostCookieJar(OkHttpClient.Builder builder) {
